[ADMIN] : [CHANGE] Edquiz Landing Page from Image to be Text && [CHANGE] Style JSX to be external file && [CHANGE] Page to be component and use with layout
