+++ eduroom_frontend/components/template/generalnonav.js
@@ -1,0 +1,11 @@
+import React, { Fragment } from 'react'
+import Header from '../layouts/header'
+const GeneralNonav = (props) => {
+  return (
+    <Fragment>
+      <Header />
+      <main>{props.children}</main>
+    </Fragment>
+  )
+}
+export default GeneralNonav
