+++ eduroom_frontend/styles/edqiz/edquizText.js
@@ -1,0 +1,15 @@
+import css from 'styled-jsx/css'
+export default css`
+  .text-title {
+    font-family: 'Quicksand', sans-serif;
+  }
+  .navy-text {
+    color: #3d467f;
+  }
+  .purple-text {
+    color: #a880f7;
+  }
+  .pink-text {
+    color: #f39ac4;
+  }
+`
